+++ api/controllers/toasterController.js
@@ -1,9 +1,13 @@
 var JsonDB = require('node-json-db');
 var db = new JsonDB("roastyDB", true, true);
 
+var utils = require('./controllerUtils.js');
+
 exports.getAllToasters = function(req, res) {
     try {
-        checkKey(req.header);
+        if (!utils.checkAPIKey(req)) {
+            throw "Invalid API Key";
+        }
         var toasters = db.getData('/toasters');
         res.send(toasters);
     } catch(e) {
@@ -15,6 +19,9 @@ exports.getAllToasters = function(req, res) {
 
 exports.addToaster = function(req, res) {
     try {
+        if (!utils.checkAPIKey(req)) {
+            throw "Invalid API Key";
+        }
         for (var i in req.body) {
             db.push('/toasters[]', req.body[i]);
         }
@@ -30,6 +37,9 @@ exports.addToaster = function(req, res) {
 
 exports.getToaster = function(req, res) {
     try {
+        if (!utils.checkAPIKey(req)) {
+            throw "Invalid API Key";
+        }
         var toaster = db.getData('/toasters['+req.params.toaster_id+']')
         res.send(toaster);
     } catch(e) {
@@ -41,6 +51,9 @@ exports.getToaster = function(req, res) {
 
 exports.updateToaster = function(req, res) {
     try {
+        if (!utils.checkAPIKey(req)) {
+            throw "Invalid API Key";
+        }
         var toaster = db.push('/toasters['+req.params.toaster_id+']', req.body[0]);
         res.send({
             'result': 'success'
@@ -54,6 +67,9 @@ exports.updateToaster = function(req, res) {
 
 exports.deleteToaster = function(req, res) {
     try {
+        if (!utils.checkAPIKey(req)) {
+            throw "Invalid API Key";
+        }
         db.delete('/toasters['+req.params.toaster_id+']');
         res.send({
             'result': 'success'
@@ -65,7 +81,3 @@ exports.deleteToaster = function(req, res) {
     }
 
 };
-
-function checkKey(header) {
-    console.log(header);
-}
